+++ src/LaravelJsonApi.php
@@ -24,6 +24,7 @@ use InvalidArgumentException;
 use LaravelJsonApi\Core\Auth\AuthorizerResolver;
 use LaravelJsonApi\Core\Query\Custom\ExtendedQueryParameters;
 use LaravelJsonApi\Eloquent\Resources\Relation;
+use LaravelJsonApi\Laravel\Http\Requests\RequestResolver;
 
 final class LaravelJsonApi
 {
@@ -53,6 +54,80 @@ final class LaravelJsonApi
     public static function defaultAuthorizer(string $authorizerClass): self
     {
         AuthorizerResolver::useDefault($authorizerClass);
+
+        return new self();
+    }
+
+    /**
+     * Register a HTTP query class for the supplied resource type or types.
+     *
+     * @param string $queryClass
+     * @param $resourceTypes
+     * @return $this
+     */
+    public static function registerQuery(string $queryClass, $resourceTypes): self
+    {
+        foreach (Arr::wrap($resourceTypes) as $resourceType) {
+            RequestResolver::register(RequestResolver::QUERY, $resourceType, $queryClass);
+        }
+
+        return new self();
+    }
+
+    /**
+     * Set the default query class implementation.
+     *
+     * @param string $queryClass
+     * @return static
+     */
+    public static function defaultQuery(string $queryClass): self
+    {
+        RequestResolver::useDefault(RequestResolver::QUERY, $queryClass);
+
+        return new self();
+    }
+
+    /**
+     * Register a HTTP collection query class for the supplied resource type or types.
+     *
+     * @param string $queryClass
+     * @param $resourceTypes
+     * @return $this
+     */
+    public static function registerCollectionQuery(string $queryClass, $resourceTypes): self
+    {
+        foreach (Arr::wrap($resourceTypes) as $resourceType) {
+            RequestResolver::register(RequestResolver::COLLECTION_QUERY, $resourceType, $queryClass);
+        }
+
+        return new self();
+    }
+
+    /**
+     * Set the default collection query class implementation.
+     *
+     * @param string $queryClass
+     * @return static
+     */
+    public static function defaultCollectionQuery(string $queryClass): self
+    {
+        RequestResolver::useDefault(RequestResolver::COLLECTION_QUERY, $queryClass);
+
+        return new self();
+    }
+
+    /**
+     * Register a HTTP request class for the supplied resource type or types.
+     *
+     * @param string $queryClass
+     * @param $resourceTypes
+     * @return $this
+     */
+    public static function registerRequest(string $queryClass, $resourceTypes): self
+    {
+        foreach (Arr::wrap($resourceTypes) as $resourceType) {
+            RequestResolver::register(RequestResolver::REQUEST, $resourceType, $queryClass);
+        }
 
         return new self();
     }